Created application factory and moved view implementation into a general blueprint
